refactor(react-dogfood): drop Ask AI Agent modal, invite directly
Clicking "Ask AI Agent" in the call controls now POSTs to
api.demo.visionagents.ai immediately instead of opening an interim
modal. The button shows "Inviting…" while the request is in flight
and the floating status panel still appears once the agent joins.

Also removes the now-unused modal component, modal SCSS, and the
translation keys that were only used by the modal copy.
