+++ sample-apps/react/react-dogfood/components/VisionAgent/index.ts
@@ -1,3 +1,2 @@
 export { AskAIAgentButton } from './AskAIAgentButton';
-export { AskAIAgentModal } from './AskAIAgentModal';
 export { AIAgentStatusPanel } from './AIAgentStatusPanel';
