+++ sample-apps/react/react-dogfood/style/VisionAgent/index.scss
@@ -29,145 +29,8 @@ $ai-gradient: linear-gradient(135deg, #005fff 0%, #7f5af0 100%);
     opacity: 0.6;
   }
 
-  &--active {
-    box-shadow: 0 0 0 2px rgba(127, 90, 240, 0.45);
-  }
-
   &__label {
     line-height: 1;
-  }
-}
-
-.rd__ask-ai-modal__backdrop {
-  position: fixed;
-  inset: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 100;
-  padding: var(--str-video__spacing-lg);
-}
-
-.rd__ask-ai-modal {
-  width: 100%;
-  max-width: 480px;
-  padding: var(--str-video__spacing-lg);
-  background-color: var(--str-video__base-color7);
-  border-radius: var(--str-video__border-radius-lg);
-  display: flex;
-  flex-direction: column;
-  gap: var(--str-video__spacing-md);
-  box-shadow: 0 24px 48px rgba(0, 0, 0, 0.4);
-
-  &__header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: var(--str-video__spacing-md);
-  }
-
-  &__heading {
-    margin: 0;
-    font-size: var(--str-video__font-size-lg);
-    font-weight: 600;
-  }
-
-  &__description {
-    margin: 0;
-    font-size: var(--str-video__font-size-sm);
-    color: var(--str-video__text-color2);
-  }
-
-  &__cards {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: var(--str-video__spacing-md);
-  }
-
-  &__card {
-    display: flex;
-    flex-direction: column;
-    gap: var(--str-video__spacing-xs);
-    padding: var(--str-video__spacing-md);
-    background-color: var(--str-video__base-color6);
-    border: 1px solid var(--str-video__base-color5);
-    border-radius: var(--str-video__border-radius-md);
-  }
-
-  &__card-icon {
-    width: 32px;
-    height: 32px;
-    border-radius: var(--str-video__border-radius-sm);
-    background-color: var(--str-video__base-color5);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  &__card-title {
-    font-size: var(--str-video__font-size-sm);
-    font-weight: 600;
-  }
-
-  &__card-description {
-    font-size: var(--str-video__font-size-xs);
-    color: var(--str-video__text-color2);
-  }
-
-  &__actions {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: var(--str-video__spacing-md);
-    margin-top: var(--str-video__spacing-xs);
-  }
-
-  &__invite {
-    appearance: none;
-    border: none;
-    color: #fff;
-    background: $ai-gradient;
-    padding: 0.5rem 1rem;
-    border-radius: 999px;
-    cursor: pointer;
-    font-size: var(--str-video__font-size-sm);
-    font-weight: 600;
-    transition: box-shadow 0.15s ease;
-
-    &:hover:not(:disabled) {
-      box-shadow: 0 4px 14px rgba(127, 90, 240, 0.35);
-    }
-
-    &:disabled {
-      cursor: not-allowed;
-      opacity: 0.6;
-    }
-  }
-
-  &__build {
-    display: inline-flex;
-    align-items: center;
-    gap: var(--str-video__spacing-xs);
-    padding: 0.5rem 1rem;
-    border-radius: 999px;
-    color: #fff;
-    background: $ai-gradient;
-    text-decoration: none;
-    font-size: var(--str-video__font-size-sm);
-    font-weight: 600;
-
-    &:hover {
-      box-shadow: 0 4px 14px rgba(127, 90, 240, 0.35);
-    }
-  }
-
-  &__footer,
-  &__powered-by {
-    margin: 0;
-    text-align: center;
-    font-size: var(--str-video__font-size-xs);
-    color: var(--str-video__text-color2);
   }
 }
