+++ sample-apps/react/react-dogfood/translations/en.json
@@ -35,15 +35,7 @@
   "device-selection/system": "Use system default",
   "device-selection/recent": "Most recently used",
   "Ask AI Agent": "Ask AI Agent",
-  "Ask an AI agent": "Ask an AI agent",
-  "Add an AI agent to this call…": "Add an AI agent to this call. Ask it anything about Stream — it's built with Vision Agents, our framework for real-time AI agents.",
-  "Talk to it": "Talk to it",
-  "Real-time voice conversation": "Real-time voice conversation",
-  "Knows Stream": "Knows Stream",
-  "Ask about our SDKs, APIs, and products": "Ask about our SDKs, APIs, and products",
-  "Invite agent": "Invite agent",
   "Build your own AI Agent": "Build your own AI Agent",
-  "The AI agent leaves the call after 5 minutes.": "The AI agent leaves the call after 5 minutes.",
   "Powered by Vision Agents": "Powered by Vision Agents",
   "AI agent in call": "AI agent in call",
   "Remove": "Remove",
